Fixed bug introduced in 5c1a961afee913e675c92a7ecc7d15ab308f61d8
This commit short-circuited broadcastability of shapes for zero-size
rhs in __setitem__.

Refix array API test failure without introducing this regression and
reuse _manipulation_functions._broadcast_strides as suggested by
@ndgrigorian
